+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master (Unreleased)
 
+* Fix `RSpec/ImplicitSubject` false positive within alias provided by this-party gems. ([@inkstak][])
 * Fix incorrect path suggested by `RSpec/FilePath` cop when second argument contains spaces. ([@tejasbubane][])
 * Fix autocorrect for EmptyLineSeparation. ([@johnny-miyake][])
 
@@ -698,3 +699,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@M-Yamashita01]: https://github.com/M-Yamashita01
 [@luke-hill]: https://github.com/luke-hill
 [@johnny-miyake]: https://github.com/johnny-miyake
+[@inkstak]: https://github.com/inkstak